Rename packages to com.otilm and bump deps
Rebrand codebase by renaming base package from com.czertainly to com.otilm across main and test sources. Update pom.xml: change parent groupId to com.otilm and parent version to 1.4.1, set artifact name/description to 'scheduler', bump com.otilm:interfaces to 2.18.1-SNAPSHOT and update the Tomcat override comment. Adjust imports and references accordingly and update logback logger name to com.otilm.
