+++ pom.xml
@@ -3,23 +3,23 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
-        <groupId>com.czertainly</groupId>
+        <groupId>com.otilm</groupId>
         <artifactId>dependencies</artifactId>
-        <version>1.4.0</version>
+        <version>1.4.1</version>
     </parent>
 
     <artifactId>scheduler</artifactId>
     <version>1.1.0</version>
-    <name>CZERTAINLY-Scheduler</name>
+    <name>scheduler</name>
 
-    <description>CZERTAINLY-Scheduler</description>
+    <description>scheduler</description>
 
     <properties>
         <java.version>21</java.version>
         <c3p0.version>0.12.0</c3p0.version>
         <mchange-commons-java.version>0.4.0</mchange-commons-java.version>
         <!-- Temporary override for CVE-2026-41293 and CVE-2026-43512.
-          Remove after the parent com.czertainly:dependencies upgrades to a Spring Boot release that ships Tomcat 10.1.55+
+          Remove after the parent com.otilm:dependencies upgrades to a Spring Boot release that ships Tomcat 10.1.55+
           (expected with Spring Boot 3.5.15). Then delete this override. -->
         <tomcat.version>10.1.55</tomcat.version>
 
@@ -30,11 +30,11 @@
 
     <dependencies>
 
-        <!-- CZERTAINLY -->
+        <!-- ILM -->
         <dependency>
-            <groupId>com.czertainly</groupId>
+            <groupId>com.otilm</groupId>
             <artifactId>interfaces</artifactId>
-            <version>2.18.0</version>
+            <version>2.18.1-SNAPSHOT</version>
             <exclusions>
                 <exclusion>
                     <groupId>org.springdoc</groupId>
