--- src/main/java/com/czertainly/scheduler/Application.java
@@ -1,15 +1,0 @@
-package com.czertainly.scheduler;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.amqp.RabbitAutoConfiguration;
-import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
-
-@SpringBootApplication(exclude = {RabbitAutoConfiguration.class})
-@ConfigurationPropertiesScan
-public class Application {
-
-    public static void main(String[] args) {
-        SpringApplication.run(Application.class, args);
-    }
-}
