--- src/main/java/com/czertainly/scheduler/constants/JobConstants.java
@@ -1,11 +1,0 @@
-package com.czertainly.scheduler.constants;
-
-public class JobConstants {
-
-    public static final String CLASS_TOBE_EXECUTED = "JOBCLASS";
-
-    public static final String GROUP_NAME = "czertainly";
-
-    public static final String JOB_TRIGGER_SUFFIX = "_trigger";
-
-}
